factor out vim.{opt,g}
