+++ lua/getchoo/globals.lua
@@ -1,0 +1,2 @@
+vim.g.mapleader = ","
+vim.g.do_filetype_lua = 1
